Fixed RandomAccessStream to allow files >2GB; tweaked readers to accomodate this change.

+++ loci/formats/in/BMPReader.java
@@ -58,7 +58,7 @@ public class BMPReader extends FormatReader {
   protected int compression;
 
   /** Offset to image data. */
-  private int global;
+  private long global;
 
   // -- Constructor --
